A little bit of refactor
